@@ -1,6 +1,6 @@
 {
   "name": "@keiser/echip-webusb",
-  "version": "1.1.0",
+  "version": "1.3.3",
   "description": "Keiser Air eChip WebUSB Library",
   "keywords": [
     "keiser",
@@ -30,23 +30,25 @@
     "push": "npm run build && cd ./dist && npm publish"
   },
   "devDependencies": {
-    "@types/node": "^12.11.1",
+    "@types/node": "^12.12.14",
     "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.0.4",
+    "copy-webpack-plugin": "^5.0.5",
     "create-file-webpack": "^1.0.2",
+    "dts-bundle": "^0.7.3",
     "html-webpack-plugin": "^3.2.0",
-    "ts-loader": "^6.2.0",
-    "tslint": "^5.20.0",
-    "tslint-config-standard": "^8.0.1",
-    "typescript": "^3.6.4",
+    "ts-loader": "^6.2.1",
+    "tslint": "^5.20.1",
+    "tslint-config-standard": "^9.0.0",
+    "typescript": "^3.7.2",
     "vue": "^2.6.10",
     "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0",
     "write-json-webpack-plugin": "^1.1.0"
   },
   "dependencies": {
     "async-mutex": "^0.1.4",
+    "cockatiel": "^0.1.1",
     "crc": "^3.8.0"
   }
 }
