@@ -1,10 +1,23 @@
-import Logger from './logger'
-import OWDevice from './owDevice'
+
 import Chip from './chip'
+import { Logger } from './logger'
+import { OWDevice } from './owDevice'
 import { EChipBuilder, EChipParser, EChipObject, MachineObject } from './echipLib'
 import { Listener, Disposable } from './typedEvent'
+import { TimeoutStrategy, Policy } from 'cockatiel'
 
-export default class EChip extends Chip {
+const TIMEOUT_INTERVAL = 20000
+const RETRY_ATTEMPTS = 2
+
+const timeoutPolicy = Policy.timeout(TIMEOUT_INTERVAL, TimeoutStrategy.Cooperative)
+const retryPolicy = Policy.handleAll().retry().attempts(RETRY_ATTEMPTS)
+
+const invalidResultGenerator = () => Promise.resolve({ machineData: {}, rawData: [], validStructure: false } as EChipObject)
+const compareResults = (srcData: Array<Uint8Array>, resData: Array<Uint8Array>) => {
+  return srcData.every((page, pageIndex) => page.every((byte, index) => byte === resData[pageIndex][index]))
+}
+
+export class EChip extends Chip {
   private owDevice: OWDevice
   private data: Promise<EChipObject>
 
@@ -21,31 +34,62 @@ export default class EChip extends Chip {
 
   async clearData () {
     let newData = EChipBuilder({})
+    await retryPolicy.execute(async () => {
+      await timeoutPolicy.execute(async () => this.performClearData(newData))
+    })
+  }
+
+  private async performClearData (newData: Uint8Array[]) {
     try {
       let oldData = (await this.data).rawData
       await this.owDevice.keyWriteDiff(this.echipId, newData, oldData, false)
     } catch (error) {
-      await this.owDevice.keyWriteAll(this.echipId, newData, false)
+      try {
+        await this.owDevice.keyWriteAll(this.echipId, newData, false)
+      } catch (error) {
+        this.data = invalidResultGenerator()
+        throw error
+      }
     }
-    this.data = new Promise(r => r(EChipParser(newData)))
-    // await (this.data = this.loadData())
+    await (this.data = this.loadData())
+    const resultsMatch = compareResults(newData, (await this.data).rawData)
+    if (!resultsMatch) {
+      throw new Error('Write was unsuccesful. Data targets do not match.')
+    }
   }
 
-  async setData (machines: {[index: string]: MachineObject}) {
+  async setData (machines: { [index: string]: MachineObject }) {
     let newData = EChipBuilder(machines)
+    await retryPolicy.execute(async () => {
+      await timeoutPolicy.execute(async () => this.performSetData(newData))
+    })
+  }
+
+  async performSetData (newData: Uint8Array[]) {
     let oldData = (await this.data).rawData
-    await this.owDevice.keyWriteDiff(this.echipId, newData, oldData, false)
-    this.data = new Promise(r => r(EChipParser(newData)))
-    // await (this.data = this.loadData())
+    try {
+      await this.owDevice.keyWriteDiff(this.echipId, newData, oldData, false)
+    } catch (error) {
+      this.data = invalidResultGenerator()
+      throw error
+    }
+    await (this.data = this.loadData())
+    const resultsMatch = compareResults(newData, (await this.data).rawData)
+    if (!resultsMatch) {
+      throw new Error('Write was unsuccesful. Data targets do not match.')
+    }
   }
 
   private async loadData () {
-    let raw = await this.owDevice.keyReadAll(this.echipId, false)
-    let echipData = EChipParser(raw)
-    console.log(raw)
-    if (!echipData.validStructure) {
-      Logger.warn('Invalid Data Structure')
-    }
-    return echipData
+    return retryPolicy.execute(async () => {
+      return timeoutPolicy.execute(async () => {
+        let raw = await this.owDevice.keyReadAll(this.echipId, false)
+        let echipData = EChipParser(raw)
+        if (!echipData.validStructure) {
+          throw new Error('Invalid data structure.')
+        }
+        return echipData
+      })
+    })
   }
 }
