@@ -1,6 +1,6 @@
 import { nodeToWeb } from './transform'
 
-export default class USBDevice {
+export class USBDevice {
   private readonly vendorId: number
   private readonly productId: number
   protected connectedDevices: Array<WebUSBDevice> = []
@@ -9,10 +9,16 @@ export default class USBDevice {
     this.vendorId = vendorId
     this.productId = productId
 
-    navigator.usb.addEventListener('connect', event => { this.attached(event as USBConnectionEvent) })
-    navigator.usb.addEventListener('disconnect', event => { this.detached(event as USBConnectionEvent) })
+    if (typeof navigator.usb === 'undefined') {
+      if (typeof window.node_usb === 'undefined') {
+        throw new Error('Web-USB not supported in this browser')
+      }
+    } else {
+      navigator.usb.addEventListener('connect', event => { void this.attached(event as USBConnectionEvent) })
+      navigator.usb.addEventListener('disconnect', event => { void this.detached(event as USBConnectionEvent) })
+    }
 
-    this.checkDevices()
+    void this.checkDevices()
   }
 
   get isConnected () {
@@ -24,8 +30,12 @@ export default class USBDevice {
   }
 
   private async requestPermission () {
-    if (this.checkNodeDevices()) {
+    if (await this.checkNodeDevices()) {
       return
+    }
+
+    if (typeof navigator.usb === 'undefined') {
+      throw new Error('Web-USB not supported in this browser')
     }
 
     let device
@@ -40,26 +50,28 @@ export default class USBDevice {
       throw new Error('USB Device permission denied.')
     }
     if (device) {
-      this.connected(device)
+      await this.connected(device)
     }
   }
 
   private async checkDevices () {
-    if (this.checkNodeDevices()) {
+    if (await this.checkNodeDevices()) {
       return
     }
 
+    if (typeof navigator.usb === 'undefined') {
+      throw new Error('Web-USB not supported in this browser')
+    }
+
     let devices = await navigator.usb.getDevices()
-    devices.some(device => {
+    for (let device of devices) {
       if (this.matchesTarget(device)) {
-        this.connected(device)
-        return true
+        await this.connected(device)
       }
-      return
-    })
+    }
   }
 
-  private checkNodeDevices () {
+  private async checkNodeDevices () {
     if (typeof window.node_usb === 'undefined') {
       return false
     }
@@ -71,22 +83,22 @@ export default class USBDevice {
 
     if (nodeUsbDevice) {
       const webUsbDevice = nodeToWeb(nodeUsbDevice)
-      this.connected(webUsbDevice)
+      await this.connected(webUsbDevice)
       return true
     }
 
     return false
   }
 
-  private attached (event: USBConnectionEvent) {
+  private async attached (event: USBConnectionEvent) {
     if (event.device && this.matchesTarget(event.device)) {
-      this.connected(event.device)
+      await this.connected(event.device)
     }
   }
 
-  private detached (event: USBConnectionEvent) {
+  private async detached (event: USBConnectionEvent) {
     if (event.device && this.matchesTarget(event.device) && this.isConnectedDevices(event.device)) {
-      this.disconnected(event.device)
+      await this.disconnected(event.device)
     }
   }
 
